move listing-fields.js into root src/ directory
